Revert "Merge branch 'sqldatespan' of https://github.com/Wraith2/corefx into sqldatespan"
This reverts commit b0ce21a2a1d68115c8c0941c1e3355c5eaaf56a2, reversing
changes made to 4e686ea76a308126970b073951ac2edf0cf52be6.

+++ src/Common/src/CoreLib/System/NotFiniteNumberException.cs
@@ -55,13 +55,13 @@ namespace System
 
         protected NotFiniteNumberException(SerializationInfo info, StreamingContext context) : base(info, context)
         {
-            _offendingNumber = info.GetDouble("OffendingNumber"); // Do not rename (binary serialization)
+            _offendingNumber = info.GetInt32("OffendingNumber");
         }
 
         public override void GetObjectData(SerializationInfo info, StreamingContext context)
         {
             base.GetObjectData(info, context);
-            info.AddValue("OffendingNumber", _offendingNumber, typeof(double)); // Do not rename (binary serialization)
+            info.AddValue("OffendingNumber", _offendingNumber, typeof(int));
         }
 
         public double OffendingNumber
